% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{Transforms a factor variable into a numeric variable}
\usage{
\method{as.numeric}{factor}(x)
}
\arguments{
\item{x}{The factor variable.}
}
\value{
A numeric variable.
}
\description{
This function transforms a factor variable into a numeric variable.
}
