% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{barplots}
\alias{barplots}
\title{Draws a barplot according to factors}
\usage{
barplots(data, x, fill, title, xlab, lx, namelegend, labellegend)
}
\arguments{
\item{data}{A data frame or a data frame extension (e.g. a tibble).}

\item{x}{The name of the first factor variable.}

\item{fill}{The name of the second factor variable.}

\item{title}{The text for the title.}

\item{xlab}{The text for the x axis.}

\item{namelegend}{The text for the legend.}

\item{labellegend}{The text for the legend labels.}
}
\value{
Barplot.
}
\description{
This function plots a barplot for a factor variable according
to the factors of another variable in the dataset.
}
