% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boxplots}
\alias{boxplots}
\title{Draws boxplots according to factors}
\usage{
boxplots(data, y, fill, title, ylab, namelegend, labellegend)
}
\arguments{
\item{data}{A data frame or a data frame extension (e.g. a tibble).}

\item{y}{The name of the numeric variable to be used.}

\item{fill}{The name of the factor variable.}

\item{title}{The text for the title.}

\item{ylab}{The text for the y axis.}

\item{namelegend}{The text for the legend.}

\item{labellegend}{The text for the legend labels.}
}
\value{
Boxplots.
}
\description{
This function plots a boxplot for a numeric variable for each factor
of a variable in the database.
}
