% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap-package.R
\docType{package}
\name{datamap-package}
\alias{datamap}
\alias{datamap-package}
\title{datamap: Interactive Visualization of Data Matrices}
\description{
A Shiny application for visualizing data matrices with heatmaps, PCA, and t-SNE. Allows users to upload, transform, and visualize their data interactively.
}
\author{
\strong{Maintainer}: Steven Ge \email{Xijin.Ge@sdstate.edu}

}
\keyword{internal}
