% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{datamap_resource}
\alias{datamap_resource}
\title{Get DataMap package resource path}
\usage{
datamap_resource(...)
}
\arguments{
\item{...}{Path components to append}
}
\value{
Full path to the resource within the package
}
\description{
Get DataMap package resource path
}
\keyword{internal}
