% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_upload.R
\name{file_upload_server}
\alias{file_upload_server}
\title{Server function for file upload module}
\usage{
file_upload_server(id)
}
\arguments{
\item{id}{The module namespace id}
}
\value{
A list of reactive values: data, data_loaded, has_rownames, and reproducible_code
}
\description{
Server function for file upload module
}
