% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_upload.R
\name{file_upload_ui}
\alias{file_upload_ui}
\title{UI function for file upload module}
\usage{
file_upload_ui(id)
}
\arguments{
\item{id}{The module namespace id}
}
\value{
A tagList of UI elements
}
\description{
UI function for file upload module
}
