% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findNA}
\alias{findNA}
\title{Find NAs in a dataset}
\usage{
findNA(dataframe)
}
\arguments{
\item{dataframe}{A data frame, data frame extension (e.g. a tibble) or a matrix}
}
\value{
A list with the number of missing values and
the columns where they are found.
}
\description{
This function finds the number of missing values in a dataset
and the columns where they are found.
}
