% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hists}
\alias{hists}
\title{Mean, standard deviation and histogram}
\usage{
hists(data, x, name, color)
}
\arguments{
\item{data}{A data frame or a data frame extension (e.g. a tibble).}

\item{x}{The name of the numeric column to be used.}

\item{name}{The name of the variable to be plotted on the histogram.}

\item{color}{The color of the histogram bars.}
}
\value{
The mean and the standard deviation of the variable.
}
\description{
This function calculates the mean and standard deviation of a numeric
column in a dataset and returns its histogram.
}
