% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modelsML}
\alias{modelsML}
\title{Compares different machine learning methods}
\usage{
modelsML(data, response, models, n = 10, k = 3, seed = 11, tune = 3)
}
\arguments{
\item{data}{A data frame or a data frame extension (e.g. a tibble).}

\item{response}{The name of the response variable.}

\item{models}{A character vector with the names of the methods to be compared.
The names must be as in names(caret::getModelInfo()).}

\item{n}{The number of folds to be used for cross validation.}

\item{k}{The number of complete sets of folds.}

\item{seed}{A single value.}

\item{tune}{An integer number of default values for each parameter to be tested.}
}
\value{
A list with the performance results of the methods.
}
\description{
This function compares different machine learning methods
through cross validation.
}
