% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mutate_column}
\alias{mutate_column}
\title{Transform the columns of a dataset}
\usage{
mutate_column(data, return_factors = T)
}
\arguments{
\item{data}{A data frame or a data frame extension (e.g. a tibble).}

\item{return_factors}{Should the factor variables be returned? Defaults to TRUE.}
}
\value{
A dataframe with the alterations.
}
\description{
This function transform the columns of a dataset into factors or numerics
depending on the number of different observations found (<=10 factor, otherwise
numeric).
}
