% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess}
\alias{preprocess}
\title{Transform the columns of a dataset}
\usage{
preprocess(data, factors, cut = 0.75)
}
\arguments{
\item{data}{A data frame or a data frame extension (e.g. a tibble).}

\item{factors}{The names of the factor variables of the dataset.}

\item{cut}{The threshold that determines high correlation.}
}
\value{
A dataframe with the alterations.
}
\description{
This function turns factors into dummies, checks for near zero covariates
in the dataset and checks for highly correlated variables.
}
