% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_column_annotations}
\alias{process_column_annotations}
\title{Process Column Annotations for Heatmap}
\usage{
process_column_annotations(main_data_cols, annotation_df, selected_annotations)
}
\arguments{
\item{main_data_cols}{A character vector of column names corresponding to the main dataset.}

\item{annotation_df}{A data frame containing annotation information where rows correspond to different annotation types
and columns represent samples.}

\item{selected_annotations}{A vector indicating the rows (by their names or indices) in annotation_df to be selected.}
}
\value{
A data frame with the main dataset's column names as its row names and the selected annotation types as its column names.
If there are no annotations selected, no common samples between datasets, or if all annotations fail to process, the function returns NULL.
}
\description{
This function processes column annotations by aligning selected annotation rows with the main dataset's columns.
It attempts to safely subset and merge annotation data based on common sample names and handles errors during processing.
}
