% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_app.R
\name{run_app}
\alias{run_app}
\title{Run the DataMap Shiny application}
\usage{
run_app()
}
\value{
A Shiny application object
}
\description{
Run the DataMap Shiny application
}
\examples{
if (interactive()) {
  run_app()
}
}
