% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{train_test}
\alias{train_test}
\title{Splits the data}
\usage{
train_test(data, response, pct = 0.8)
}
\arguments{
\item{data}{A data frame or a data frame extension (e.g. a tibble).}

\item{response}{The name of the response variable.}

\item{pct}{The percentage of data that goes to training.}
}
\value{
A list with the partitions.
}
\description{
Splits the data into training, training label, test and test label.
}
