% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_transform.R
\name{transform_server}
\alias{transform_server}
\title{Server function for preprocessing module}
\usage{
transform_server(id, data)
}
\arguments{
\item{id}{The module namespace id}

\item{data}{Reactive data frame to process}
}
\value{
A list with the processed data reactive and reproducible code
}
\description{
Server function for preprocessing module
}
