% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_transform.R
\name{transform_ui}
\alias{transform_ui}
\title{UI function for preprocessing module button}
\usage{
transform_ui(id)
}
\arguments{
\item{id}{The module namespace id}
}
\value{
A button that triggers the preprocessing modal
}
\description{
UI function for preprocessing module button
}
