% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{XGboost.class}
\alias{XGboost.class}
\title{Fits an XGBoost model to the data}
\usage{
XGboost.class(train, label, n = 10, k = 3, tune = 6, seed = 11)
}
\arguments{
\item{train}{The training data without the label.}

\item{label}{The training label.}

\item{n}{The number of folds to be used for cross validation.}

\item{k}{The number of complete sets of folds.}

\item{tune}{An integer number of default values for each parameter to be tested.}

\item{seed}{A single value.}
}
\value{
A list containing the final model, the probabilities estimated by the model,
the AUC, the threshold selected through the ROC curve and the confusion matrix.
}
\description{
This function fits an eXtreme Gradient Boosting model to the data
with the main parameters chosen through cross validation
}
