% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{datamap_resource}
\alias{datamap_resource}
\title{Find resource files for datamap}
\usage{
datamap_resource(path)
}
\arguments{
\item{path}{Path to the resource within the package}
}
\value{
The full path to the resource file
}
\description{
Find resource files for datamap
}
\keyword{internal}
